requirements.txt
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
idna==2.8
isort==4.3.17
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==2.3.1
python-dotenv==0.10.3
requests==2.21.0
six==1.12.0
typed-ast==1.3.5
urllib3==1.24.2
wrapt==1.11.1
